feat: implementing router with aliases @ to remove ../ file path
