+++ frontend/package.json
@@ -13,9 +13,11 @@
     "primeicons": "^7.0.0",
     "primevue": "^4.2.3",
     "vue": "^3.5.12",
+    "vue-router": "^4.5.0",
     "vuedraggable": "^4.1.0"
   },
   "devDependencies": {
+    "@types/node": "^22.10.1",
     "@vitejs/plugin-vue": "^5.1.4",
     "typescript": "~5.6.2",
     "vite": "^5.4.10",
