+++ frontend/src/main.ts
@@ -4,7 +4,15 @@ import App from './App.vue'
 import PrimeVue from 'primevue/config';
 import Aura from '@primevue/themes/aura';
 import 'primeicons/primeicons.css';
+import { createRouter, createWebHistory } from 'vue-router';
+import Home from '@/views/Home.vue';
 
+const router = createRouter({
+    history: createWebHistory(),
+    routes:[
+        {path: '/', name: "Home", component: Home}
+    ]
+})
 
 const app = createApp(App);
 app.use(PrimeVue, {
@@ -12,4 +20,6 @@ app.use(PrimeVue, {
         preset: Aura
     }
 });
+
+app.use(router)
 app.mount('#app')
