+++ frontend/vite.config.ts
@@ -1,7 +1,13 @@
 import { defineConfig } from 'vite'
 import vue from '@vitejs/plugin-vue'
+import {resolve} from "path";
 
 // https://vite.dev/config/
 export default defineConfig({
+  resolve : {
+    alias : {
+      "@": resolve(__dirname, "src")
+    }
+  },
   plugins: [vue()],
 })
